requirements.txt
httpx~=0.24.1
pandas~=1.5.3
joblib~=1.3.2
pytest~=7.4.2
ttkbootstrap~=1.10.1
matplotlib~=3.7.3
Pillow~=10.1.0
scikit-learn~=1.2.2
pdoc3~=0.10.0